refactor: Rename mdigits components to activity
