+++ lib/src/components/activity/task_step.dart
