+++ lib/src/components/response/response_controller.dart
@@ -1,14 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:mdigits/src/data.dart';
-import 'package:mdigits/src/components/mdigits/mdigits_controller.dart';
+import 'package:mdigits/src/components/activity/activity_controller.dart';
 import 'package:mdigits/src/components/stim/stim_controller.dart';
 import 'package:mdigits/src/components/config/app_config.dart';
 
 /// Manage responses
 class ResponseController extends GetxController {
   TextEditingController textController = TextEditingController();
-  final MDigitsController _mDigits = Get.find();
+  final ActivityController _activityController = Get.find();
   final Data _data = Get.find();
   final AppConfig _config = Get.find();
   final StimController _stim = Get.find();
@@ -34,6 +34,6 @@ class ResponseController extends GetxController {
   }
 
   void toNextScreen() {
-    _mDigits.endRestOrContinueTask();
+    _activityController.endRestOrContinueTask();
   }
 }
