+++ lib/src/components/rest/rest_controller.dart
@@ -1,13 +1,13 @@
 import 'package:get/get.dart';
-import 'package:mdigits/src/components/mdigits/mdigits_controller.dart';
+import 'package:mdigits/src/components/activity/activity_controller.dart';
 
 class RestController extends GetxController {
-  final MDigitsController _mDigitsController = Get.find();
+  final ActivityController _activityController = Get.find();
 
   /// Goes to next screen after 500ms ISI
   void toNextScreen() {
     Future.delayed(const Duration(milliseconds: 500), () {
-      _mDigitsController.endRestOrContinueTask();
+      _activityController.endRestOrContinueTask();
     });
   }
 }
